requirements.txt
amqp==5.0.6
asgiref==3.3.4
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.2.2
django-extensions==3.1.3
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drfdocs2==0.0.14
kombu==5.0.2
Markdown==3.3.4
prompt-toolkit==3.0.18
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
